Removed resolve-url-loader again and replaced bit by a link /cc @tobyS
The resolve-url-loader does not work together with the
MiniCssExtractPlugin which is *required* to extract the production
assets. (In theory it should, but it just doens't). I use the old (ugly)
pathes again and created a lin `catwalk/` -> `paas/catwalk/` to make it
possible to resolve those paths. This is an *ugly hack* but makes it
possible to actually deploy something.

We *must* also always adapt and test the production webpack config if we
change the development config!