Added custom yarn-link folder
This allows us to have yarn links that work within the VM and on the host machine.

The Symlink in `paas/.yarn-links` is a work around that we might want to discuss.

The problem is that the `yarn link frontastic-common --link-folder ../.yarn-links` is being executed within the apollo-de directory and within the paas/catwalk directory.

Maybe it would even make sense to set —link-folder to .yarn-links and have symlinks from both apollo_de and paas/catwalk.

I am open for discussions on that topic.

(project merge back 2019-05-06)