Grid: windowWidth/gridWidth should not be passed to dom elements
This should remove the following error:

Warning: React does not recognize the `gridWidth` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `gridwidth` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
(project merge back 2019-05-06)