Implemented locale based content fetching for Contentful.
So far we expected Contentful to return all translations for a given
content object. This is not always the case, depending on the Contentful
configuration. We agreed with several ContentApi implementors to include
the locale in the fetch request in order to fetch only the desired
translated data from the CMS back-end. This also reduces usage efforts
for the developer in the fron-end, API traffic and enables the CMS
fall-back mechanisms for translations to hook in (as we already do it
for commerce).