Add migration to update collation
This ensures the collation and the character set is updated in all developer VMs. It does not update the locale of the app_* tables. When they are newly created they are now created with the correct collation. But existing tables have to be updated manually (using paas/integration/bin/migrateCollation)