Remove unused streams and find stream count definitions
We are using an include list for known count properties for now. If we
do not find any count property for a stream we default to no count
limits.

If there is one stream usage with a count definition and one without
this will limit the stream count for now and thus might be a BC break.