BCB: Source basic Catwalk Symfony configuration automatically.
Before we imported the base configuration in the project configuration
files. This is not necessary anymore which makes project config more
independent and disallows to remove our base config.

Note that this is a tiny BC break: The import statements for the base
config needs to be removed from project config files in
*/config/config*.yml.