Create store lazily
We only create the store during hydration and not in a export-singleton
anymore. This changes the app initialization logic slightly. Most
importantly the store is not created any more more as soon as some
depending component in the webpack tree depends on the `app/app` but
only during hydration.

This ensures that ComponentInjector.addReducer() always works as
expected.