Adds possibility to add a custom classname to nodes
Currently confusing in this process is that we have two
files that appear to be default schema settings
(paas/libraries/common/src/js/configuration/defaultSchemas/nodeConfiguration.json
and
paas/libraries/common/test/php/SpecificationBundle/Domain/_fixtures/node/valid/default_schema.json)
as well as the version with the current settings for each project
(demo_english/src/js/nodeConfiguration.json) which also has to be
uploaded into the schema editor in backstage. This process feels
a bit redundant and might be a good place to start for some
optimizations.