Ensure CSS modules do not break styling during SSR
We still extract (dummy) CSS during SSR so we get the proper class names
out of the CSS modules. Otherwise we are left with `undefined` class
names in the HTML tree and the React hydration does not always fix them
which causes an unclean first rendered version of the page.