Stop enforcing error statuscode for all exceptions
Before this change, whenever we had an exception in our stack that was handled by the symfony exception listeners,
the response had a 500 status code, even if some exception listener handled the exception.
This caused that a customer that did redirects in an exception handler still received 500s on those responses instead of the 30x, which resulted in bad SEO experience :(

With this code, only if no other exception listener would create a response, we still set an error status code.