feat!: Catr/Wishlist Loaders separate methods for getContinuously().
Previously there was only a get() method for these loaders which
triggered the continuous reloading. This method was renamed to
getContinuously() and a method get() was introduced which fetches only
once and returns a promise.

This should be more intuitive to project developers and should not
interfere with existing projects even though it is a BC break.