fix: removes resize event handlers from grid
the current browser dimensions are now being pulled
from the redux store, instead of registering new
event listeners for every grid instance.

Also, in cell.jsx a small code hack was added that
temporarily solves odd behaviour on SSR when
browser dimensions are not known yet.

The idea is that the client and SSR versions do not
differ in any way, which is impossible to do if the
JS relies on browser values for rendering. So for
future versions, we should try to remove the dependency
on JS-based browser information and try to solve it
with plain CSS.