fix: Run yarn install lso for no-paas customers
yarn install has only be run for catwalk until now, which was fine,
because we are suing workspaces and thus it was sufficient to run yarn
install only once. But since no-paas customers have no catwalk this
meant that during build `yarn install` was not run at all for them (only
during provisioning). This caused old libraries to be used in the build.