feat: moves notifications to more global place
This is a work in progress and generally bad practice,
to grab something from boost theme into catwalk.
Proper solution is currently being brainstormed,
but the overall point is to remove all the little
instances of notifications in all kinds of child
components. We have context, state and other tools
for that.