fix: return error code in debug error controller
Talked about this with @hco and @schaumiii. I changed it to return 200 because Chronext calls `/__error` after switching the locale on an error page and our JavaScript code does not render a node on an error status code. But this controller should never be used on production. So changing it back to 500.