fix!: Switch to default symfony error handling
We deliberately do not catch exceptions anymore, because this is how symfony wants it.
Symfony instead registers exception & error handlers as well as a shutdown handler,
which will then in turn trigger the frontastic error listener.

If we would catch Throwable in here, the symfony error handling will not work,
as it relies on the exception bubbling up. Exception handling will work as expected,
because for exceptions symfony does a catch(Exception), but not for Throwable,
which means Errors are excluded.

And yes, this means that there’s a risk that some errors/exceptions are not handled properly, but we are taking the risk deliberatly.