fix!: properly arrange import order for encrypted vaults
Fixes the order of the import to comply with what we have in the docs here: https://docs.frontastic.cloud/docs/catwalk-configuration-for-advanced-users

We had a bug in the sourcing of the project.ymls with the encrypted ones, that we always sourced the crypt at latest which was not complying with the docs. Behavior before was:

*Production*
1. project.yml
2. project.yml.crypt

*Staging*
1. project.yml
2. project.yml.staging
3. project.yml.crypt
[4. project.yml.staging.crypt - if present]

So the project.yml.crypt also got sourced and taken into account on staging if there was no project.yml.staging.crypt.
This behavior is changed now and complies to the docs now!

*Production*
1. project.yml
2. project.yml.crypt

*Staging*
1. project.yml
2. project.yml.staging
[3. project.yml.staging.crypt - if present]

This also means that if you relied on that bug, you'll need to check if the configuration is still working properly for you and add maybe necessary information either to a project.yml.staging.crypt or to the project.yml.staging!