fix: output proper HTML5 appData div tag
Self-closing syntax (<div />) in HTML5 is only supported for
void-elements (elements which never can contain any content). See
https://html.spec.whatwg.org/multipage/syntax.html#start-tags (6.)

Error recovery of browsers treats the <div /> as start tag, meaning that
everything after it is nested inside the div. It seems to not break anything,
but I don't see any reason to not adhere to the standard.