fix: Actually make our stateless requests stateless.
The Symfony 'stateless' attribute in security.yml does *not* ensure a
session is not created. If you ask `$request->hasSession()` at any point
in your request, the session will be started regarless of this setting.

We now use a custom route attribute `_stateless` and check for this one
in the places where we access the session and it is not mandatory
(Context request listener).