fix: send DynamicPageContext to backend extension runner
When the dynamic-page-handler was requested from the extension runner, only the
API Request was sent. The context was left out as a TODO. This change adds the
DynamicPageContext to the request so that the extension runner can pass it to
the extension (which is not yet done).

The DynamicPageContext had to be created by the FromFrontasticReactMapper from a
regular Context.