fix(api-hub): introduce FT NextJs specific locale resolver
The default (Frontastic React) LocaleResolver uses mostly the session to
determine the locale to use and falls back to the default one when no session is
set. That also leads to not being able to set the locale through the request URL
when there is no session. As in Frontastic NextJs there is never a session, the
URL cannot be used to set the locale for Frontastic NextJs.

This change creates a Frontastic NextJs specific locale resolver which doesn't
use the session at all but looks at the request URL (locale parameter) and if
that is not there the Accept header of the browser before falling back to the
projects default locale.