feat(FP-1252): allow field type dataSource in schemas
Optionally allow field type dataSource and specifying the type with the property dataSourceType in schemas. This change requires updates to the studio as well as to the API hub. Customers need to update their code before they can use this feature. For this reason this feature is opt-in. So customers are free to use stream/streamType until they update their code.

This change does not work for Frontastic Classic. Uploading schemas using the dataSource field type is not forbidden in Classic but it will probably break in Catwalk.