feat(FP-2420): Install path-to-regexp into catwalk explicitly
The recent module changes lead to a different resolution order for path-to-regexp which in turn leads to react-router using an extremely outdated version of this package. That breaks routing. Unfortunately the only quick fix is to bump path-to-regexp up in resolution order by making it a primary dependency of paas/catwalk.